added: finished

+++ put_task_in_progress.py
@@ -1,17 +1,25 @@
 import json
+import os
+from datetime import datetime
+
 
 def put_task_in_progress(task_to_put_in_progress_id):
-    with open('tasks.json', 'r') as f:
-        tasks = json.load(f)
-        if not isinstance(tasks, list):
-            tasks = [tasks]
+    if os.path.exists('tasks.json'):
+        with open('tasks.json', 'r') as f:
+            tasks = json.load(f)
+            if not isinstance(tasks, list):
+                tasks = [tasks]
 
-    task_to_finish = next((t for t in tasks if t['id'] == task_to_put_in_progress_id), None)
+        task_to_finish = next((t for t in tasks if t['id'] == task_to_put_in_progress_id), None)
 
-    if task_to_finish:
-        task_to_finish['status'] = "In progress"
-        tasks.sort(key=lambda task: task['id'])
-        with open('tasks.json', 'w') as f:
-            json.dump(tasks, f, indent=4)
+        if task_to_finish:
+            task_to_finish['status'] = "in progress"
+            task_to_finish['updated_at'] = datetime.now()
+            tasks.sort(key=lambda task: task['id'])
+            with open('tasks.json', 'w') as f:
+                json.dump(tasks, f, indent=4)
+            print(f"Task: {task_to_finish['description']} was marked as in progress")
+        else:
+            print(f"No task with id {task_to_put_in_progress_id} was found.")
     else:
-        print(f"No task with id {task_to_put_in_progress_id} was found.")
+        print("No tasks found. make some tasks")
